+++ protocols/gossipsub/src/lib.rs
@@ -138,6 +138,7 @@
 pub mod error;
 pub mod protocol;
 
+mod backoff;
 mod behaviour;
 mod config;
 mod gossip_promises;
